+++ writereader.go
@@ -37,7 +37,7 @@ import (
 // On Plan 9, atomic creation of a new file is supported, but atomic replacement
 // of an existing file is not. If the destination exists, WriteReader returns an
 // error matching errors.ErrUnsupported and leaves the destination unchanged.
-// To work around this issue, use the WithAllowNonAtomicReplace option.
+// To work around this issue, use the WithNonAtomicReplace option.
 func WriteReader(name string, r io.Reader, perm os.FileMode, opts ...Option) (err error) { //nolint:funlen,gocyclo
 	fopts := Options{
 		flags:        os.O_CREATE | os.O_WRONLY | os.O_TRUNC,
@@ -132,7 +132,7 @@ func WriteReader(name string, r io.Reader, perm os.FileMode, opts ...Option) (er
 		return writeError(name, err)
 	}
 
-	err = Rename(tempFileName, name, WithAllowNonAtomicReplace(fopts.allowNonAtomicReplace))
+	err = Rename(tempFileName, name, WithNonAtomicReplace(fopts.nonAtomicReplace))
 	if err != nil {
 		err = fmt.Errorf("cannot rename to '%v': %w", tempFileName, err)
 		return writeError(name, err)